Add phpstan.neon configuration file and update phpstan command in Makefile

+++ src/Bookshop/Catalog/Application/Service/Book/Remove/RemoveBookService.php
@@ -5,6 +5,7 @@ namespace Bookshop\Catalog\Application\Service\Book\Remove;
 use Bookshop\Catalog\Application\Service\Book\Remove\RemoveBookRequest;
 use Bookshop\Catalog\Domain\Event\Book\BookRemovedEvent;
 use Bookshop\Catalog\Domain\Event\DomainEventPublisher;
+use Bookshop\Catalog\Domain\Model\Book\BookDoesNotExistException;
 use Bookshop\Catalog\Domain\Model\Book\BookId;
 use Bookshop\Catalog\Domain\Model\Book\BookRepository;
 
@@ -23,11 +24,17 @@ class RemoveBookService
 
     public function execute(RemoveBookRequest $request): void
     {
-        $bookId = new BookId($request->bookId());
+        $bookId = new BookId($request->id());
+
         $book = $this->bookRepository->ofBookId($bookId);
+
+        if (null === $book) {
+            throw new BookDoesNotExistException();
+        }
+
         $this->bookRepository->remove($book);
 
-        $event = new BookRemovedEvent($book->bookId());
+        $event = new BookRemovedEvent($bookId);
         $this->domainEventPublisher->publish($event);
     }
 }
