+++ src/Bookshop/Catalog/Application/Service/Genre/Remove/RemoveGenreService.php
@@ -5,7 +5,6 @@ namespace Bookshop\Catalog\Application\Service\Genre\Remove;
 use Bookshop\Catalog\Domain\Model\Genre\GenreDoesNotExistException;
 use Bookshop\Catalog\Domain\Model\Genre\GenreId;
 use Bookshop\Catalog\Domain\Model\Genre\GenreRepository;
-use Exception;
 
 class RemoveGenreService
 {
@@ -19,7 +18,13 @@ class RemoveGenreService
     public function execute(RemoveGenreRequest $request): void
     {
         $genreId = new GenreId($request->genreId());
+
         $genre = $this->service->ofGenreId($genreId);
+
+        if ($genre === null) {
+            throw new GenreDoesNotExistException();
+        }
+
         $this->service->remove($genre);
     }
 }
